SEparated the JS script
